# Conflicts:
#	lib/app/CMakeLists.txt

+++ lib/app/src/app/component/chemistry/bond.cpp
@@ -6,11 +6,11 @@ namespace VTX::App::Component::Chemistry
 	size_t Bond::getIndexFirstAtom() const { return _moleculePtr->_bondAtomPairIds[ _internalIndex * 2 ]; }
 	void   Bond::setIndexFirstAtom( const size_t p_atomIndex )
 	{
-		_moleculePtr->_bondAtomPairIds[ _internalIndex * 2 ] = p_atomIndex;
+		_moleculePtr->_bondAtomPairIds[ _internalIndex * 2 ] = uint( p_atomIndex );
 	}
 	size_t Bond::getIndexSecondAtom() const { return _moleculePtr->_bondAtomPairIds[ _internalIndex * 2 + 1 ]; }
 	void   Bond::setIndexSecondAtom( const size_t p_atomIndex )
 	{
-		_moleculePtr->_bondAtomPairIds[ _internalIndex * 2 + 1 ] = p_atomIndex;
+		_moleculePtr->_bondAtomPairIds[ _internalIndex * 2 + 1 ] = uint( p_atomIndex );
 	}
 } // namespace VTX::App::Component::Chemistry
